# Biological_app/server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bioserver api",
    "description": "This is Bioserver api document\n",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/grorge/bioserver/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "https://bioapp-backend.yikuo.dev/",
      "description": "Bioserver"
    },
    {
      "url": "http://localhost:8080",
      "description": "Bioserver local"
    }
  ],
...
  "paths": {
...
    "/api/data_bird": {
      "get": {
        "tags": [
          "user_bird"
        ],
        "summary": "取得使用者上傳過的資料",
        "description": "取得使用者曾經上傳過的資料，st代表開始位置，從0開始計算，回傳num筆，如果使用者上傳編號未到st+num筆則只回傳到使用者目前最大筆數。",
        "operationId": "get_data_bird",
        "parameters": [
          {
            "name": "st",
            "in": "query",
            "description": "資料順序開始位置",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "num",
            "in": "query",
            "description": "取得資料數量",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "user_bird"
        ],
        "summary": "更新使用者的鳥類資料",
        "description": "修改之前上傳過的資料，id為要修改資料的編號",
        "operationId": "put_data_bird",
        "requestBody": {
          "description": "you should provide new password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/put_bird_data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "example": {
                  "status": "success"
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "user_bird"
        ],
        "summary": "上傳使用者的鳥類資料",
        "description": "取得使用者上傳過的資料",
        "operationId": "post_data_bird",
        "requestBody": {
          "description": "you should provide new password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/post_bird_data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "id": 0
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "user_bird"
        ],
        "summary": "刪除使用者上傳的鳥類資料",
        "description": "刪除使用者曾經上傳過的資料，id為要刪除資料的編號",
        "operationId": "delete_data_bird",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "要刪除資料編號",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "invalid",
            "in": "query",
            "description": "資料是否違法",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "example": {
                  "status": "success"
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/near_data_bird": {
      "get": {
        "tags": [
          "user_bird"
        ],
        "summary": "取得目前最近鳥類資料",
        "description": "取得目前最近鳥類資料，st代表開始位置，從0開始計算，回傳num筆，如果使用者上傳編號未到st+num筆則只回傳到使用者目前最大筆數，lon代表使用者經度，lat代表使用者緯度。",
        "operationId": "get_near_data_bird",
        "parameters": [
          {
            "name": "st",
            "in": "query",
            "description": "資料順序開始位置",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "num",
            "in": "query",
            "description": "取得資料數量",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "lat",
            "in": "query",
            "description": "以某緯度做最近資料搜查，需要和lon一起填寫",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "lon",
            "in": "query",
            "description": "以某經度做最近資料搜查，需要和lat一起填寫",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "指定某個id的資料",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "以某項參數做排序,可接授\"id\", \"name\", \"description\", \"lon\", \"lat\", \"time\"",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "default": "id"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "DESC或是ASC",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string",
              "default": "DESC"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "description": "關鍵字搜尋",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "upload",
            "in": "query",
            "description": "是否上鏈",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_1"
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/money": {
      "get": {
        "tags": [
          "other"
        ],
        "summary": "取得使用者身上點數",
        "description": "取得使用者身上點數",
        "operationId": "get_money",
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_3"
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "other"
        ],
        "summary": "給予其他使用者金錢",
        "description": "給予其他使用者金錢",
        "operationId": "post_money",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api_money_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "response_ERROR": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "default": "error"
          },
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "get_bird_data": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "lon": {
            "type": "integer"
          },
          "lat": {
            "type": "integer"
          },
          "upload": {
            "type": "boolean"
          },
          "time": {
            "type": "integer"
          }
        }
      },
      "post_bird_data": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "lon": {
            "type": "integer"
          },
          "lat": {
            "type": "integer"
          },
          "time": {
            "type": "integer"
          }
        }
      },
      "put_bird_data": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "inline_response_200": {
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/get_bird_data"
            }
          },
          "len": {
            "type": "integer"
          }
        }
      },
      "inline_response_200_1": {
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/get_bird_data"
            }
          },
          "len": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "inline_response_200_3": {
        "properties": {
          "status": {
            "type": "string",
            "default": "success"
          },
          "money": {
            "type": "integer",
            "default": 0
          }
        }
      },
      "api_money_body": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "money": {
            "type": "integer"
          }
        }
      },
      "inline_response_200_4": {
        "properties": {
          "status": {
            "type": "string",
            "default": "success"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}